+++ CHANGES.md
@@ -23,6 +23,7 @@ master (in development)
   - `chef-foodcritic` handles relative paths correctly now [GH-556]
   - Global Flycheck Mode enables Flycheck Mode even if there is no syntax
     checker for the buffer yet [GH-568]
+  - Remove `global-flycheck-mode` from customization interface [GH-595]
 
 - Bug fixes: